requirements.txt
llama-index-core>=0.10.11.post1
llama-index-llms-gemini>=0.1.4
llama-index-embeddings-gemini>=0.1.4
llama-index-readers-file>=0.1.0
google-generativeai>=0.3.2
pypdf>=3.17.1
python-dotenv>=1.0.0
streamlit>=1.29.0
nltk>=3.8.1